# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TinyClaw API",
    "version": "1",
    "description": "HTTP API for the TinyClaw personal AI assistant."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4310",
      "description": "Local dev server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "SystemStatusResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": {}
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/system/status": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "System status",
        "operationId": "getSystemStatus",
        "responses": {
          "200": {
            "description": "Server and automation worker status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}